// app/metais-admin/src/api/generated/cmdb-swagger.ts
export type RoleType = (typeof RoleType)[keyof typeof RoleType]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RoleType = {
    SYSTEM: 'SYSTEM',
    NON_SYSTEM: 'NON_SYSTEM',
} as const

export interface Role {
    assignedGroup?: string
    category?: string
    description?: string
    name?: string
    type?: RoleType
    uuid?: string
    weight?: number
}

export interface RoleParticipantUI {
    configurationItemUi?: ConfigurationItemUi
    gid?: string
    owner?: string
    role?: Role
}
// ...
export interface RelationshipUi {
    attributes?: AttributeUi[]
    endUuid?: string
    metaAttributes?: MetaAttributesUi
    owner?: string
    startUuid?: string
    type?: string
    uuid?: string
}
// ...
export interface PaginationUi {
    page?: number
    perPage?: number
    totalPages?: number
    totaltems?: number
}
// ...
export interface NeighboursResultUi {
    neighbourPairs?: NeighbourPairUi[]
    pagination?: PaginationUi
}
// ...
export interface NeighbourSetUi {
    fromNodes?: NeighboursResultUi
    toNodes?: NeighboursResultUi
}

export interface NeighbourPairUi {
    configurationItem?: ConfigurationItemUi
    relationship?: RelationshipUi
}

export interface MissingAttributesHolderUi {
    lastModification?: string
    missingAttributes?: string[]
}

export interface MetaIsCodesListUi {
    metaIsCodes?: string[]
}

export interface MetaAttributesUi {
    createdAt?: string
    createdBy?: string
    lastModifiedAt?: string
    lastModifiedBy?: string
    owner?: string
    state?: string
}
// ...
export type ConfigurationItemUiAttributes = { [key: string]: any }

export interface ConfigurationItemUi {
    attributes?: ConfigurationItemUiAttributes
    metaAttributes?: MetaAttributesUi
    owner?: string
    type?: string
    uuid?: string
}

export interface ConfigurationItemsListUi {
    configurationItemSet?: ConfigurationItemUi[]
}

export interface ConfigurationItemSetUi {
    configurationItemSet?: ConfigurationItemUi[]
    incidentRelationshipSet?: RelationshipUi[]
    pagination?: PaginationUi
}

export interface ConfigurationItemNeighbourSetUi {
    fromCiSet?: ConfigurationItemUi[]
    toCiSet?: ConfigurationItemUi[]
}

export interface ConfigurationItemInvalidateUi {
    attributes?: AttributeUi[]
    invalidateReason?: InvalidateReason
    metaAttributes?: MetaAttributesUi
    owner?: string
    type?: string
    uuid?: string
}

export interface CiWithRelsUi {
    ci?: ConfigurationItemUi
    rels?: RelationshipUi[]
}

export interface CiWithRelsResultUi {
    ciWithRels?: CiWithRelsUi[]
    pagination?: PaginationUi
}
// ...
export interface AttributeUi {
    name?: string
    value?: AttributeUiValue
}